// nautilus-ionic/src/app/api/model/association.ts
/**
 * nautilus-central-service API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface Association { 
    id?: number;
    name?: string;
    address?: string;
    email?: string;
    phone?: string;
    website?: string;
    remarks?: string;
    img64?: Blob;
}